Rearrange code, run goimports.
